macrursors-sel: Better secondary selection
* macrursors-sel.el (macrursors-sel--select, macrursors-sel-types,
macrursors-sel--filter-cursors, macrursors-sel-cycle,
macrursors-sel-map, macrursors-sel-clear,
macrursors-sel-mark-from-isearch): It is now possible to interact
with the secondary selection using the keyboard. Calling
`macrursors-sel-cycle' will cycle through secondary selections
covering objects listed in `macrursors-sel-types'. This command
activates a transient keymap so it can be called again with a
single key. Currently only defun, list and line objects are
available. When doing so, macro cursors outside the secondary
selection will be discarded. `macrursors-sel-mark-from-isearch'
will create cursors at occurrences of the current isearch string.
This is bound in `isearch-mode-map'.