macrursors-sel: Update cursors when selection enlarges
* macrursors.el (macrursors-mark-all-lines,
macrursors--defun-mark-all, macrursors-mark-previous-instance-of,
macrursors-mark-next-instance-of,
macrursors--mark-all-instances-of): Update cursor locations when
enlarging the region. This works by storing the current search
string or syntactic unit (word/sexp/line etc) in a new variable
and re-running the search when the secondary selection changes.
Note that the search only re-runs when the selection grows in
either direction around point.

(macrursors--instance): Buffer-local variable holding the last
thing we searched for to create cursors.

* macrursors-sel.el (macrursors-sel-mark-from-isearch,
macrursors-sel-cycle): When the secondary selection changes, check
if the search needs to be re-run and do it.

* macrursors-expand.el (macrursors-sel-expand): Ditto.