macrursors: New prefix-arg behavior when marking next/previous
* macrursors.el: (macrursors--mark-next-instance-of,
macrursors-mark-next-instance-of,
macrursors--mark-previous-instance-of,
macrursors-mark-previous-instance-of,
macrursors-mark-next-from-isearch,
macrursors-mark-previous-from-isearch, macrursors-mark-next-line,
macrursors-mark-previous-line): Add a new behavior when calling
macrursors-mark-next-* and macrursors-mark-previous-* commands
with zero prefix-arg.  They will now mark all occurrences of the
thing to the end of or beginning of the buffer.

This includes marking lines, marking from isearch and the generic
mark-next command (which marks regions, symbols or lines).